Extract a Glooko API client
This adds a client for the Glooko REST API.  The code was extracted from
`lib/sources/glooko/index.js`, and refactored into a decoupled module
that can be tested in isolation from the rest of the project's
components.

This adds Mocha as a dev dependency, and configures npm to run it via
the `npm test` command.

This also adds js-beautify and a script that can be run via `npm
run-script fmt` to apply the [cgm-remote-monitor style guide][1] to the
two new `api.js` files.

[1]: https://github.com/nightscout/cgm-remote-monitor/blob/15.0.3/CONTRIBUTING.md#style-guide